move files into expose folder
