+++ bin/release.js
@@ -1,3 +1,9 @@
 #!/usr/bin/env node
 
 console.log("release");
+
+// first webpack build. This will set assets/dist
+// then copy in the build dir what needs to be exposed
+// the set proper URL in layouts AND in specs.yml
+// bundle for fast syncing (zip)
+// upload to S3
