+++ lib/expose/utils/replace_assets_host.js
