+++ lib/expose/watch_file_system.js
